[DROOLS-6236] merge 2 subsequent accumulates when possible
