+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/constraints/ConstraintEvaluator.java
@@ -71,10 +71,6 @@ public class ConstraintEvaluator {
         return constraint;
     }
 
-    public InnerEvaluator getInnerEvaluator() {
-        return innerEvaluator;
-    }
-
     private InnerEvaluator initInnerEvaluator(Declaration patternDeclaration) {
         if (isTemporal()) {
             setPatternDeclaration( patternDeclaration );
